Source code is reorganized to match original Python library.
