+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+## Changelog
+
+### 1.4.0
+
+- Library became CommonJS compatible.
+- Source code is reorganized to match [original Python library](https://pypi.python.org/pypi/sgp4/).
