+++ src/propagation.js
@@ -1,0 +1,3 @@
+export { default as propagate } from './propagation/propagate';
+export { default as sgp4 } from './propagation/sgp4';
+export { default as gstime } from './propagation/gstime';
